requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8

# File Processing
PyPDF2==3.0.1
lxml==4.9.3

# HTTP Requests
requests==2.31.0

# LLM Integration
openai==1.3.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment Variables
python-dotenv==1.0.0
